requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
langchain_community
pdf2image
moviepy
openai-whisper
ffmpeg-python
pydub